Add files via upload
Suppose you want to deposit a certain amount of money into a savings account and leave it alone to draw interest for the next 10 years. At the end of 10 years, you would like to have $10,000 in the account. How much do you need to deposit today to make that happen? You can use the following formula to find out:

P = F / (1 + r) ^n

where P is the present value, or the amount that you need to deposit today, F is the future value that you want in the account (in this case, F is $10,000), r is the annual interest rate, and n is the number of years that you plan to let the money sit in the account.

 

Program Algorithm:

1. Get the designed future value

2. Get the annual interest rate

3. Get the number of years that the money will sit in the account

4. Calculate the amount that will have to be deposited.

5. Display the result of the calculation in step 4.